requirements.txt
pytz
Django==4.0.2
djangorestframework==3.11.0
django-debug-toolbar
drf-yasg